Remove duplication and clean up
